docs: update cascader
