--- docs/demo/cascader/section-forbidclear.jsx
@@ -1,165 +1,0 @@
-import React from 'react'
-import DocViewer from '../../../libs/doc-viewer'
-import Cascader from '../../../components/cascader'
-const prefix = 'section-forbidclear'
-const rightOptions = ['禁止清空', '不禁止']
-const code = [
-  {
-    code: `import React from 'react'
-import Cascader from '@hi-ui/hiui/es/cascader'\n
-class Demo extends React.Component {
-  constructor () {
-    super()
-    this.state = {
-      options: [
-        {
-          id: '手机',
-          content: '手机',
-          children: [
-            {
-              id: '小米',
-              content: '小米',
-              children: [
-                {
-                  id: '小米3',
-                  content: '小米3'
-                },
-                {
-                  id: '小米4',
-                  content: '小米4'
-                },
-              ]
-            },
-            {
-              id: '红米',
-              content: '红米',
-              children: [
-                {
-                  id: '红米3',
-                  content: '红米3'
-                },
-                {
-                  id: '红米4',
-                  content: '红米4'
-                }
-              ]
-            }
-          ]
-        },
-        {
-          id: '电视',
-          content: '电视',
-          children: [
-            {
-              id: '小米电视4A',
-              content: '小米电视4A'
-            },
-            {
-              id: '小米电视4C',
-              content: '小米电视4C'
-            }
-          ]
-        }
-      ]
-    }
-  }
-  render(){
-    return(
-      <Cascader
-        id={['手机', '红米', '红米4']}
-        onChange={(id)=>{
-          console.log('on change', id)
-        }}
-        clearable={false}
-        data={this.state.options}
-        style={{ width: 240 }}
-      />
-    )
-  }
-}`,
-    opt: ['禁止清空']
-  }, {
-    code: `import React from 'react'
-import Cascader from '@hi-ui/hiui/es/cascader'\n
-class Demo extends React.Component {
-  constructor () {
-    super()
-    this.state = {
-      options: [
-        {
-          id: '手机',
-          content: '手机',
-          children: [
-            {
-              id: '小米',
-              content: '小米',
-              children: [
-                {
-                  id: '小米3',
-                  content: '小米3'
-                },
-                {
-                  id: '小米4',
-                  content: '小米4'
-                },
-              ]
-            },
-            {
-              id: '红米',
-              content: '红米',
-              children: [
-                {
-                  id: '红米3',
-                  content: '红米3'
-                },
-                {
-                  id: '红米4',
-                  content: '红米4'
-                }
-              ]
-            }
-          ]
-        },
-        {
-          id: '电视',
-          content: '电视',
-          children: [
-            {
-              id: '小米电视4A',
-              content: '小米电视4A'
-            },
-            {
-              id: '小米电视4C',
-              content: '小米电视4C'
-            }
-          ]
-        }
-      ]
-    }
-  }
-  render(){
-    return(
-      <Cascader
-        id={['手机', '红米', '红米4']}
-        onChange={(id)=>{
-          console.log('on change', id)
-        }}
-        data={this.state.options}
-        style={{ width: 240 }}
-      />
-    )
-  }
-}`,
-    opt: ['不禁止']
-  }
-]
-
-const DemoBasic = () => (
-  <DocViewer
-    code={code}
-    scope={{ Cascader }}
-    prefix={prefix}
-    rightOptions={rightOptions}
-  />
-)
-export default DemoBasic
