+++ docs/demo/select/section-type.jsx
@@ -2,9 +2,107 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Select from '../../../components/select'
 const prefix = 'alert-autoClose'
-const rightOptions = ['正常', '受控']
+const rightOptions = ['基础', '受控', '带默认值', '可清空', '禁用']
 const desc = '展示从多个收起的备选项中选出的一个选项'
 const code = [
+  {
+    code: `import React from 'react'
+import Select from '@hi-ui/hiui/es/select'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      value: '3',
+      singleList: [
+        { title:'电视', id:'3', disabled: true },
+        { title:'手机', id:'2' },
+        { title:'笔记本', id:'4', disabled: true },
+        { title:'生活周边', id:'5' },
+        { title:'办公', id:'6' },
+      ]
+    }
+  }
+
+  render () {
+    const { value, singleList } = this.state
+    return (
+      <Select
+        type='single'
+        clearable={false}
+        style={{ width: 200 }}
+        data={singleList}
+      />
+    )
+  }
+}`,
+    opt: ['基础']
+  },
+  {
+    code: `import React from 'react'
+import Select from '@hi-ui/hiui/es/select'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      value: '3',
+      singleList: [
+        { title:'电视', id:'3', disabled: true },
+        { title:'手机', id:'2' },
+        { title:'笔记本', id:'4', disabled: true },
+        { title:'生活周边', id:'5' },
+        { title:'办公', id:'6' },
+      ]
+    }
+  }
+
+  render () {
+    const { value, singleList } = this.state
+    return (
+      <Select
+        type='single'
+        clearable={false}
+        style={{ width: 200 }}
+        data={singleList}
+        disabled
+      />
+    )
+  }
+}`,
+    opt: ['禁用']
+  },
+  {
+    code: `import React from 'react'
+import Select from '@hi-ui/hiui/es/select'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      value: '3',
+      singleList: [
+        { title:'电视', id:'3', disabled: true },
+        { title:'手机', id:'2' },
+        { title:'笔记本', id:'4', disabled: true },
+        { title:'生活周边', id:'5' },
+        { title:'办公', id:'6' },
+      ]
+    }
+  }
+
+  render () {
+    const { value, singleList } = this.state
+    return (
+      <Select
+        type='single'
+        clearable={false}
+        style={{ width: 200 }}
+        data={singleList}
+        clearable
+      />
+    )
+  }
+}`,
+    opt: ['可清空']
+  },
   {
     code: `import React from 'react'
 import Select from '@hi-ui/hiui/es/select'\n
@@ -36,7 +134,7 @@ class Demo extends React.Component {
     )
   }
 }`,
-    opt: ['正常']
+    opt: ['带默认值']
   },
   {
     code: `import React from 'react'
