+++ docs/zh-CN/components/cascader.mdx
@@ -1,18 +1,21 @@
 # Cascader 级联选择
 
+一种接收数据的容器，为用户提供选择一部分数据的能力
+
+## 何时使用
+
+- 需要从大量的离散型数据中选择一部分时使用
+- 备选项数量5个以上时
+- 不需要将全部备选项都展示给用户时
+
 ## 基础用法
 
 import DemoBasic from '../../demo/cascader/section-basic.jsx'
 
 <DemoBasic />
 
-## 禁止清空
 
-import DemoForbidclear from '../../demo/cascader/section-forbidclear.jsx'
-
-<DemoForbidclear />
-
-## 高级用法
+## 特殊用法
 
 import DemoAdvanced from '../../demo/cascader/section-advanced.jsx'
