+++ docs/zh-CN/components/select.mdx
@@ -16,9 +16,6 @@ import DemoType from '../../demo/select/section-type'
 
 <DemoType />
 
-## 禁止状态
-
-import DemoBan from '../../demo/select/section-ban'
 
 <DemoBan />
